xong login

+++ src/main/java/com/example/watersystem/repository/AdminUserRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AdminUserRepository extends JpaRepository<AdminUser, Long> {
+public interface AdminUserRepository extends JpaRepository<AdminUser, Integer> {
 
     Optional<AdminUser> findByUsername(String username);
