+++ src/main/java/com/example/watersystem/repository/InvoiceRepository.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
+public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
 
     List<Invoice> findByStatus(String status);
