+++ src/main/java/com/example/watersystem/repository/PaymentRepository.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 
     List<Payment> findByInvoice(Invoice invoice);
